+++ src/o365/yammer/commands.ts
@@ -1,6 +1,9 @@
 const prefix: string = 'yammer';
 
-export default {  
+export default {
   YAMMER_MESSAGE_GET: `${prefix} message get`,
-  YAMMER_NETWORK_LIST: `${prefix} network list`
+  YAMMER_MESSAGE_LIST: `${prefix} message list`,
+  YAMMER_NETWORK_LIST: `${prefix} network list`,
+  YAMMER_USER_GET: `${prefix} user get`,
+  YAMMER_USER_LIST: `${prefix} user list`
 };