+++ docs/manual/docs/about/release-notes.md
@@ -1,12 +1,33 @@
 # Release notes
 
-## v2.2.0
+## v2.3.0
+
+### New commands
+
+**SharePoint Framework**
+
+- [spfx project externalize](../cmd/spfx/project/project-externalize.md) - externalizes SharePoint Framework project dependencies [#571](https://github.com/pnp/office365-cli/issues/571)
+
+**Yammer**
+
+- [yammer message list](../cmd/yammer/message/message-list.md) - returns all accessible messages from the user's Yammer network [#1104](https://github.com/pnp/office365-cli/issues/1104)
+- [yammer user list](../cmd/yammer/user/user-list.md) - returns users from the current network [#1113](https://github.com/pnp/office365-cli/issues/1113)
+
+### Changes
+
+- added the 'Authenticate with and call the Microsoft Graph' example [#1186](https://github.com/pnp/office365-cli/issues/1186)
+- fixed the 'spo hubsite list' command [#1180](https://github.com/pnp/office365-cli/issues/1180)
+- fixed the 'spo file add' command [#1179](https://github.com/pnp/office365-cli/issues/1179)
+
+## [v2.2.0](https://github.com/pnp/office365-cli/releases/tag/v2.2.0)
 
 ### New commands
 
 **Azure Active Directory:**
 
 - [aad o365group report activitydetail](../cmd/aad/o365group/o365group-report-activitydetail.md) - get details about Office 365 Groups activity by group [#1130](https://github.com/pnp/office365-cli/issues/1130)
+- [aad o365group report activitycounts](../cmd/aad/o365group/o365group-report-activitycounts.md) - get the number of group activities across group workloads [#1159](https://github.com/pnp/office365-cli/issues/1159)
+- [aad o365group report activitygroupcounts](../cmd/aad/o365group/o365group-report-activitygroupcounts.md) - get the daily total number of groups and how many of them were active based on email conversations, Yammer posts, and SharePoint file activities [#1160](https://github.com/pnp/office365-cli/issues/1160)
 
 **Flow:**
 
@@ -28,6 +49,7 @@
 **Yammer:**
 
 - [yammer network list](../cmd/yammer/network/network-list.md) - returns a list of networks to which the current user has access [#1115](https://github.com/pnp/office365-cli/issues/1115)
+- [yammer user get](../cmd/yammer/user/user-get.md) - retrieves the current user or searches for a user by ID or e-mail [#1107](https://github.com/pnp/office365-cli/issues/1107)
 
 ### Changes
 
@@ -36,6 +58,10 @@
 - added the 'remove custom themes' example [#1137](https://github.com/pnp/office365-cli/issues/1137)
 - corrected 'aad o365group user list' alias [#1149](https://github.com/pnp/office365-cli/issues/1149)
 - updated 'spo storageentity set' docs about handling trailing slash [#1153](https://github.com/pnp/office365-cli/issues/1153)
+- updated vorpal to 1.11.7 [#1150](https://github.com/pnp/office365-cli/issues/1150)
+- added versions to deps for building docs in CI [#1157](https://github.com/pnp/office365-cli/issues/1157)
+- added the 'consent' command [#1162](https://github.com/pnp/office365-cli/issues/1162)
+- added the 'Delete custom SharePoint site designs' example [#1138](https://github.com/pnp/office365-cli/issues/1138)
 
 ## [v2.1.0](https://github.com/pnp/office365-cli/releases/tag/v2.1.0)